Remove some macros to make RDoc happier.
When generating documentation RDoc cannot look inside
of macros to extract possible constants.  This means
that the constants for domain states, storage pool states,
and volume states were missing.  While having the macros
is slightly nicer, it's not really a huge deal.  Therefore,
remove the macros so that RDoc can generate proper
documentation.

Signed-off-by: Chris Lalancette <clalance@redhat.com>