Re-implement Libvirt::open_auth with blocks.
Previously Libvirt::open_auth was relying on passing a
symbol that represented a callback ruby function to call
when credentials were needed.  While this worked, it was
not idiomatic to ruby.  The right thing to do here is to use
blocks, which is exactly what this patch does.  As a side
benefit, I think it greatly simplifies the code and simplifies
the documentation, so that is a net win.

Note that although this is an API break in the public git
repository, no release has ever been done with the previous
version of Libvirt::open_auth, so I think this is safe to do.

Signed-off-by: Chris Lalancette <clalance@redhat.com>